// src/app/core/model/contact.model.ts
export interface IContactAPI {

    _id?: string;

    sectionTitle: string;

    email: string;

    phone: string;

    github: string;

    linkedin: string;

    showEmail: boolean;

    showPhone: boolean;

    showGithub: boolean;

    showLinkedin: boolean;

}

export interface IContact {

    id?: string;

    sectionTitle: string;

    email: string;

    phone: string;

    github: string;

    linkedin: string;

    showEmail: boolean;

    showPhone: boolean;

    showGithub: boolean;

    showLinkedin: boolean;

}